style: move ts interfaces into their own files in models
