+++ bundle-es2015/models/IUpdateNode.d.ts
@@ -1,0 +1,8 @@
+import { Optional } from '../utils';
+import { Node } from './Node';
+/**
+ * A Node-like interface containing required properties to update an
+ * existing Node and all other properties are optional
+ */
+export interface IUpdateNode extends Optional<Node, 'childCount' | 'path' | 'proxy' | 'renderer'> {
+}
