+++ bundle-es2015/utils.d.ts
@@ -1,0 +1,1 @@
+export declare type Optional<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>> & Pick<Partial<T>, K>;
