+++ lib/models/index.d.ts
@@ -25,6 +25,7 @@ export * from './GroupListOptions';
 export * from './GroupUserListOptions';
 export * from './IComponentOperations';
 export * from './IContentTypeOperations';
+export * from './ICreateNode';
 export * from './IEntryOperations';
 export * from './IEventOperations';
 export * from './IGroupOperations';
@@ -34,6 +35,7 @@ export * from './IProjectOperations';
 export * from './IRedirectOperations';
 export * from './IRoleOperations';
 export * from './ISecurityOperations';
+export * from './IUpdateNode';
 export * from './IUserOperations';
 export * from './ManagementApiScope';
 export * from './Node';
