+++ lib/models/index.js
@@ -28,6 +28,7 @@ tslib_1.__exportStar(require("./GroupListOptions"), exports);
 tslib_1.__exportStar(require("./GroupUserListOptions"), exports);
 tslib_1.__exportStar(require("./IComponentOperations"), exports);
 tslib_1.__exportStar(require("./IContentTypeOperations"), exports);
+tslib_1.__exportStar(require("./ICreateNode"), exports);
 tslib_1.__exportStar(require("./IEntryOperations"), exports);
 tslib_1.__exportStar(require("./IEventOperations"), exports);
 tslib_1.__exportStar(require("./IGroupOperations"), exports);
@@ -37,6 +38,7 @@ tslib_1.__exportStar(require("./IProjectOperations"), exports);
 tslib_1.__exportStar(require("./IRedirectOperations"), exports);
 tslib_1.__exportStar(require("./IRoleOperations"), exports);
 tslib_1.__exportStar(require("./ISecurityOperations"), exports);
+tslib_1.__exportStar(require("./IUpdateNode"), exports);
 tslib_1.__exportStar(require("./IUserOperations"), exports);
 tslib_1.__exportStar(require("./ManagementApiScope"), exports);
 tslib_1.__exportStar(require("./Node"), exports);
