+++ src/models/Node.ts
@@ -1,12 +1,7 @@
 import { VersionInfo } from 'contensis-core-api';
 
-// Make specific members of an object optional and union the others
-type Optional<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>> &
-  Pick<Partial<T>, K>;
-
 type Proxy = any | null;
 type Renderer = any | null;
-
 
 /**
  * A Node as it is returned from the Management API
@@ -27,29 +22,4 @@ export interface Node {
   includeInMenu: boolean;
   version: Partial<VersionInfo>;
 }
-
-/**
- * A Node-like interface containing required properties to create a new Node
- * and all other properties are optional
- */
-export interface ICreateNode
-  extends Optional<
-    Node,
-    | 'childCount'
-    | 'id'
-    | 'includeInMenu'
-    | 'isCanonical'
-    | 'path'
-    | 'proxy'
-    | 'renderer'
-    | 'restrictedToLanguages'
-    | 'version'
-  > {}
-
-  /**
- * A Node-like interface containing required properties to update an
- * existing Node and all other properties are optional
- */
-export interface IUpdateNode
-  extends Optional<Node, 'childCount' | 'path' | 'proxy' | 'renderer'> {}
